/settle endpoint test

+++ crates/driver/src/infra/api/settle.rs
@@ -4,7 +4,7 @@ pub(super) fn route(router: axum::Router<super::State>) -> axum::Router<super::S
 
 async fn settle(
     state: axum::extract::State<super::State>,
-    axum::extract::Path(solution_id): axum::extract::Path<u32>,
+    axum::extract::Path(solution_id): axum::extract::Path<u64>,
 ) {
     // TODO Report errors instead of unwrapping
     let competition = state.competition();
